* master: (297 commits)
  Bib: Added refs and fixed one
  Doc: Fixed formatting problems for MatCreateAIJ() manual page.
  Add to PCSOR manual page information about convergence testing when using KSPRICHARDSON
  SNES ex5: Restore old output since this breaks an example I have
  Noted that some Cray XC systems will use 'aprun' instead of 'srun'. Thanks to Herr Doktor Hong Zhang for this suggestion.
  Added example configure script for Cray XC40 with KNL processors (e.g., NERSC Cori).
  Fix possible sprintf -Wformat-overflow
  PetscMalloc2-7: fix -Wint-in-bool-context for GCC 7
  SNES ex5: make boundary condition handling consistent for symmetry
  SNES ex5: use L2 norm and *_monitor_short for output
  SNES ex5: the L^2 norm scales with the l^2 norm over sqrt(N)
  SNES ex5: refactor MMS capability to avoid extreme code duplication
  Documentation fix: typo in PetscLogStageRegister manual page.
  nodes -> vertices
  rename DMNetworkGetComponentTypeOffset -> DMNetworkGetComponentKeyOffset
  rename DMNetworkGetConnectedNodes() -> DMNetworkGetConnectedVertices()
  Jacobians for adjoints for explicit methods and constant Jacobians never got initialized
  fix complex build error
  fix build failure with fortran interface
  fix compile error with cdc753b64c658d532d7813e942002e67025667a6
  ...

Conflicts:
	src/snes/examples/tutorials/ex13.c
	src/snes/examples/tutorials/ex62.c
	src/snes/examples/tutorials/ex69.c

+++ src/snes/utils/convest.c
@@ -343,7 +343,7 @@ PetscErrorCode PetscConvEstGetConvRate(PetscConvEst ce, PetscReal *alpha)
       for (f = 0; f < ce->Nf; ++f) {
         if (f > 0) {ierr = PetscPrintf(comm, ", ");CHKERRQ(ierr);}
         if (errors[f] < 1.0e-11) {ierr = PetscPrintf(comm, "< 1e-11");CHKERRQ(ierr);}
-        else                     {ierr = PetscPrintf(comm, "%g", errors[f]);CHKERRQ(ierr);}
+        else                     {ierr = PetscPrintf(comm, "%g", (double)errors[f]);CHKERRQ(ierr);}
       }
       ierr = PetscPrintf(comm, "]\n");CHKERRQ(ierr);
     }
@@ -398,7 +398,7 @@ PetscErrorCode PetscConvEstRateView(PetscConvEst ce, PetscReal alpha, PetscViewe
   ierr = PetscObjectTypeCompare((PetscObject) viewer, PETSCVIEWERASCII, &isAscii);CHKERRQ(ierr);
   if (isAscii) {
     ierr = PetscViewerASCIIAddTab(viewer, ((PetscObject) ce)->tablevel);CHKERRQ(ierr);
-    ierr = PetscViewerASCIIPrintf(viewer, "L_2 convergence rate: %0.2g\n", (double) alpha);CHKERRQ(ierr);
+    ierr = PetscViewerASCIIPrintf(viewer, "L_2 convergence rate: %g\n", (double) alpha);CHKERRQ(ierr);
     ierr = PetscViewerASCIISubtractTab(viewer, ((PetscObject) ce)->tablevel);CHKERRQ(ierr);
   }
   PetscFunctionReturn(0);
